Added the parameter -m/--mode
The -m/--mode parameter can be either "generation" or "verification". If the generation mode is selected, then Ott-IFC will generate a new mechanism from the specification given as input. If the verification mode is selected, then Ott-IFC will check an existing mechanism to identify potential errors (this part has not yet been implemented).